Refactor some component properties and fix rendering key issue
This commit tidies up some component properties across the Language, Layout, and FormattedHTMLMessage components. It particularly fixes a rendering error in the Language component by assigning a unique key to each MenuItem. The commit also switches to a more generalised title ID in the Layout component, and capitalizes the 'tagName' property for consistency in the FormattedHTMLMessage component.