Add CookieBanner component and make minor syntax adjustments
The primary change in this commit is the introduction of a CookieBanner component, which manages cookie consent. A CookieBanner import was added to Layout.js, with this component appended to the component return.

Minor code cleanup was also performed, with double to single quotation mark changes made for standardization, and syntax spacing adjusted. Import redundancies were also removed. The file package-lock.json was deleted, probably because of a switch to Yarn from NPM as the package manager.